requirements.txt
annotated-types==0.6.0
pydantic==2.6.0
pydantic_core==2.16.1
pyhumps==3.8.0
typing_extensions==4.9.0
